convert fromt int to double in bone-frame

+++ src/com/akjava/gwt/skeltalboneanimation/client/bones/BoneFrame.java
@@ -8,7 +8,7 @@ import com.google.common.collect.Lists;
 
 public class BoneFrame {
 private String boneName;
-public BoneFrame(String boneName, int x, int y, double angle) {
+public BoneFrame(String boneName, double x, double y, double angle) {
 	super();
 	this.boneName = boneName;
 	this.x = x;
@@ -32,20 +32,20 @@ public String getBoneName() {
 public void setBoneName(String boneName) {
 	this.boneName = boneName;
 }
-public int getX() {
+public double getX() {
 	return x;
 }
-public void setX(int x) {
+public void setX(double x) {
 	this.x = x;
 }
-public int getY() {
+public double getY() {
 	return y;
 }
-public void setY(int y) {
+public void setY(double y) {
 	this.y = y;
 }
-private int x;
-private int y;
+private double x;
+private double y;
 private double angle;
 public double getAngle() {
 	return angle;
